feat(cvsb-19691): moved tests to tests folder
