+++ tsconfig.json
@@ -13,7 +13,7 @@
     "esModuleInterop": true,
     "resolveJsonModule": true
   },
-  "include": ["./src/**/*.ts", "index.test.ts"],
+  "include": ["./src/**/*.ts", "tests/unit/index.test.ts", "tests/**/*.ts"],
   "exclude": [
     "node_modules/**/*",
     ".serverless/**/*",
